+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [2.5.0] - 2020-11-11
 ### Changed
 - Updated comb_spec_searcher version for faster counting
+- Added a point placement strategy to partial row and col placement packs.
 
 ## [2.4.1] - 2020-10-28
 ### Added